+++ include/rivercpp/StandardScaler.h
@@ -17,7 +17,7 @@ private:
     std::array<double, num_features> vars;
 public:
     void learn_one(const std::vector<double>& x, int y) override {
-        for (int i=0;i<x.size();i++) {
+        for (size_t i=0;i<x.size();i++) {
             counts[i] += 1;
             double old_mean = means[i];
             means[i] += (x[i] - old_mean) / counts[i];
@@ -26,7 +26,7 @@ public:
     }
     std::vector<double> transform_one(const std::vector<double>& x) override {
         std::vector<double> res(x.size(), 0.0);
-        for (int i=0;i<x.size();i++) {
+        for (size_t i=0;i<x.size();i++) {
             res[i] = vars[i] > 0.0 ? (x[i] - means[i]) / std::sqrt(vars[i]) : 0.0;
         }
         return res;
